svd2rust 0.21.0

+++ src/aes/aes_auth_length.rs
@@ -28,7 +28,7 @@ impl<'a> AUTH_LENGTH_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u32) -> &'a mut W {
-        self.w.bits = value as u32;
+        self.w.bits = value;
         self.w
     }
 }
