+++ src/aes/aes_data_in_out_0.rs
@@ -29,7 +29,7 @@ impl<'a> AES_DATA_IN_OUT_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u32) -> &'a mut W {
-        self.w.bits = value as u32;
+        self.w.bits = value;
         self.w
     }
 }
