+++ src/aes/dmac_ch0_extaddr.rs
@@ -57,7 +57,7 @@ impl<'a> ADDR_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u32) -> &'a mut W {
-        self.w.bits = value as u32;
+        self.w.bits = value;
         self.w
     }
 }
@@ -65,7 +65,7 @@ impl R {
     #[doc = "Bits 0:31 - Channel external address value When read during operation, it holds the last updated external address after being sent to the master interface."]
     #[inline(always)]
     pub fn addr(&self) -> ADDR_R {
-        ADDR_R::new(self.bits as u32)
+        ADDR_R::new(self.bits)
     }
 }
 impl W {
