+++ src/aes/hash_data_in_5.rs
@@ -28,7 +28,7 @@ impl<'a> HASH_DATA_IN_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u32) -> &'a mut W {
-        self.w.bits = value as u32;
+        self.w.bits = value;
         self.w
     }
 }
