+++ src/udma/altbase.rs
@@ -32,7 +32,7 @@ impl R {
     #[doc = "Bits 0:31 - Alternate channel address pointer This field provides the base address of the alternate channel control structures."]
     #[inline(always)]
     pub fn addr(&self) -> ADDR_R {
-        ADDR_R::new(self.bits as u32)
+        ADDR_R::new(self.bits)
     }
 }
 #[doc = "DMA alternate channel control base pointer The ALTBASE register returns the base address of the alternate channel control data. This register removes the necessity for application software to calculate the base address of the alternate channel control structures. This register cannot be read when the uDMA controller is in the reset state.\n\nThis register you can [`read`](crate::generic::Reg::read). See [API](https://docs.rs/svd2rust/#read--modify--write-api).\n\nFor information about available fields see [altbase](index.html) module"]
